backend/requirements.txt
Django==4.2.9
psycopg #add version
psycopg2 #add version
ipython
djangorestframework==3.15.1
Markdown==3.5.2
django_filter==23.5
